+++ src/codemagic/google/resources/identifiers/__init__.py
@@ -1,0 +1,2 @@
+from .app_identifier import AppIdentifier
+from .resource_identifier import ResourceIdentifier