+++ src/codemagic/tools/firebase_app_distribution/__init__.py
@@ -1,0 +1,1 @@
+from .firebase_app_distribution import FirebaseAppDistribution