+++ src/codemagic/tools/firebase_app_distribution/action_groups/firebase_action_groups.py
@@ -1,0 +1,8 @@
+from codemagic import cli
+
+
+class FirebaseActionGroups(cli.ActionGroup):
+    RELEASES = cli.ActionGroupProperties(
+        name='releases',
+        description='Manage your Firebase application releases',
+    )