+++ src/codemagic/tools/firebase_app_distribution/actions/__init__.py
@@ -1,0 +1,1 @@
+from .get_latest_build_version_action import GetLatestBuildVersionAction