+++ src/codemagic/tools/firebase_app_distribution/errors.py
@@ -1,0 +1,5 @@
+from codemagic import cli
+
+
+class FirebaseAppDistributionError(cli.CliAppException):
+    pass